code/{{cookiecutter.directory_name}}/src/layers/core/python/requirements.txt
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.5
SQLAlchemy==2.0.43
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
urllib3==2.5.0
aws_lambda_powertools==3.19.0
PyMySQL==1.1.1
